refactor: 검색 및 결과 페이지 히스토리 처리 로직 개선
