+++ src/pages/search/index/components/LoadingSpinner.tsx
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import React from 'react';
 
 const LoadingSpinner = () => {
   return (
