+++ src/pages/search/index/components/Search.tsx
@@ -1,0 +1,21 @@
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+const Search = () => {
+  const navigate = useNavigate();
+  const [startDate, setStartDate] = useState('');
+  const [endDate, setEndDate] = useState('');
+
+  const handleSearch = () => {
+    navigate(`/search?startDate=${startDate}&endDate=${endDate}`);
+  };
+  return (
+    <div>
+      <input type="date" value={startDate} onChange={(e) => setStartDate(e.target.value)} placeholder="Start Date" />
+      <input type="date" value={endDate} onChange={(e) => setEndDate(e.target.value)} placeholder="End Date" />
+      <button onClick={handleSearch}>GO</button>
+    </div>
+  );
+};
+
+export default Search;
