+++ src/pages/search/index/components/TripCards.tsx
